requirements.txt
discord.py==1.3.2
requests==2.23.0
riotwatcher==2.7.1
pytesseract==0.3.2
discord==1.0.1
Pillow==7.0.0
